third_party/blink/web_tests/external/wpt/css/css-inline/parsing/alignment-baseline-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['alignment-baseline'] = "text-bottom" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['alignment-baseline'] = "text-top" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['alignment-baseline'] = "bottom" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['alignment-baseline'] = "center" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['alignment-baseline'] = "top" should set the property value
  assert_not_equals: property should be set got disallowed value ""
Harness: the test ran to completion.
